ci: add lint gate, dependabot, aggregate status check, lockfile pinning
Four gaps from a CI review, plus the fallout from turning clippy on.

Dependabot (.github/dependabot.yml, new)
  release-crates.yml's comment promised "Dependabot will open PRs" for
  the SHA-pinned actions, but no config existed. Pinning without an
  upgrade prompt just means sitting on known-bad versions quietly.
  Covers github-actions and cargo weekly, the binding ecosystems
  (npm, bun, nuget, bundler, mix, gomod, maven) monthly, grouped so a
  quiet week is one PR rather than nine. honker-jvm/honker-kotlin have
  no CI job, so this is currently the only thing watching them.

Lint job (ci.yml, Cargo.toml, Makefile)
  Nothing ran cargo fmt or clippy. New `lint` job runs both, with a
  second clippy pass over the cfg-gated kernel-watcher/shm-fast-path
  code the default-feature pass never type-checks. Lint policy lives in
  the root [workspace.lints.clippy] table so `make lint` and CI agree.

  Turning it on surfaced 11 findings. Nine were mechanical and are
  fixed here: redundant u32 casts in cron.rs, io::Error::new(Other, _)
  -> io::Error::other, a single-arm match, an explicit copy closure, a
  collapsible if, and a manual Result::unwrap_or in the extension's
  watcher_wait. Two are design opinions (too_many_arguments on the
  published enqueue signature, type_complexity on two SQL row tuples)
  and are allowed at workspace level with rationale — refactoring a
  published API belongs in a deliberate 0.3, not a lint sweep.

Aggregate status check (ci.yml)
  The python and node matrices are built with fromJson and run a
  smaller set on pull_request than on push, so job names differ by
  event and no hand-maintained required-checks list is correct for
  both. New `ci-ok` job has one stable name, depends on every real job,
  and asserts each result is success — treating skipped as failure
  except for perf-regression, which is gated off on PRs by design.
  `if: always()` is load-bearing: a skipped required check reads as a
  pass to branch protection.

Lockfile pinning
  package-lock.json and bun.lock were gitignored, so `npm install`
  floated @napi-rs/cli ^3.0.0 and `bun install --frozen-lockfile` was
  pinning against a file that never existed. Both lockfiles are now
  committed and the node job uses `npm ci`. Every cargo invocation
  across CI and the release workflows gained --locked, including
  `cargo publish --locked` — Cargo.lock was committed but nothing
  enforced it. Only the three crates with a committed lock are
  affected; packages/honker and packages/honker-node still resolve
  freely.

Also: `apt-get update` before the C++ dep install, and an
adhoc-packages suppression for release-ruby.yml, whose two flagged
`gem install` calls install the locally built gem under test — there is
no registry to pin against.

Verified: make lint, cargo test --locked (default and experimental
features), cargo build --locked, npm ci, bun install --frozen-lockfile,
and zizmor --min-severity low across all workflows. Cargo.lock is
unchanged, confirming --locked is safe to add.

+++ honker-core/src/cron.rs
@@ -88,8 +88,8 @@ impl CronSchedule {
         self.seconds.contains(&dt.second())
             && self.minutes.contains(&dt.minute())
             && self.hours.contains(&dt.hour())
-            && self.days.contains(&(dt.day() as u32))
-            && self.months.contains(&(dt.month() as u32))
+            && self.days.contains(&dt.day())
+            && self.months.contains(&dt.month())
             && self.dows.contains(&cron_dow)
     }
 }
@@ -187,7 +187,7 @@ fn cron_day_matches(sched: &CronSchedule, dt: &NaiveDateTime) -> bool {
         Weekday::Fri => 5,
         Weekday::Sat => 6,
     };
-    sched.days.contains(&(dt.day() as u32)) && sched.dows.contains(&cron_dow)
+    sched.days.contains(&dt.day()) && sched.dows.contains(&cron_dow)
 }
 
 fn make_dt(year: i32, month: u32, day: u32, hour: u32, minute: u32, second: u32) -> NaiveDateTime {
